refactor(cli): restore release-ready command ownership
