+++ loopx/cli_commands/registry_admin.py
@@ -10,10 +10,6 @@ from ..configure_goal import configure_goal, render_configure_goal_markdown
 from ..control_plane.goals.configure_goal_service import (
     configure_goal_with_global_sync,
 )
-from ..control_plane.goals.activation_service import (
-    render_goal_activation_markdown,
-    set_goal_activation_state,
-)
 from ..file_lock import exclusive_file_lock, lock_timeout_error_fields
 from ..global_registry import (
     render_global_goal_retirement_markdown,
@@ -41,6 +37,7 @@ from ..thread_agent_binding import (
     unbind_thread_agent_in_registry,
 )
 from ..upgrade import build_upgrade_plan
+from .goal_lifecycle import handle_goal_lifecycle_command, register_goal_lifecycle_command
 from .registry_admin_configure import register_configure_goal_command
 from .registry_admin_peer import render_register_agent_markdown
 from .registry_authority import (
@@ -373,31 +370,7 @@ def loop_activation_for_goal(
 
 def register_registry_admin_commands(subparsers: argparse._SubParsersAction) -> None:
     register_configure_goal_command(subparsers)
-
-    goal_lifecycle_parser = subparsers.add_parser(
-        "goal-lifecycle",
-        help="Preview or apply a reversible Goal stop/resume transition.",
-    )
-    goal_lifecycle_parser.add_argument(
-        "--goal-id",
-        required=True,
-        help="Goal id present in the active registry.",
-    )
-    goal_lifecycle_parser.add_argument(
-        "--operation",
-        choices=("stop", "resume"),
-        required=True,
-        help="Stop automatic advancement or restore eligibility.",
-    )
-    goal_lifecycle_parser.add_argument(
-        "--reason",
-        help="Bounded owner-visible transition reason.",
-    )
-    goal_lifecycle_parser.add_argument(
-        "--execute",
-        action="store_true",
-        help="Write the source registry and verify the shared projection; preview is the default.",
-    )
+    register_goal_lifecycle_command(subparsers)
 
     register_agent_parser = subparsers.add_parser(
         "register-agent",
@@ -599,29 +572,11 @@ def handle_registry_admin_command(
         return None
 
     if args.command == "goal-lifecycle":
-        try:
-            payload = set_goal_activation_state(
-                registry_path=registry_path,
-                goal_id=args.goal_id,
-                state="stopped" if args.operation == "stop" else "active",
-                reason=args.reason,
-                runtime_root_override=args.runtime_root,
-                execute=bool(args.execute),
-            )
-        except Exception as exc:
-            payload = {
-                "ok": False,
-                "schema_version": "loopx_goal_activation_transition_v1",
-                "dry_run": not bool(args.execute),
-                "execute": bool(args.execute),
-                "goal_id": args.goal_id,
-                "changed": False,
-                "written": False,
-                "error": str(exc),
-                **lock_timeout_error_fields(exc),
-            }
-        print_payload(payload, args.format, render_goal_activation_markdown)
-        return 0 if payload.get("ok") else 1
+        return handle_goal_lifecycle_command(
+            args,
+            registry_path=registry_path,
+            print_payload=print_payload,
+        )
 
     if args.command == "configure-goal":
         try:
