+++ man/loopx.1
@@ -75,6 +75,9 @@ Render a handoff or review packet with any required evidence\-log reads.
 \fBloopx goal\-channel \-\-help\fR
 Set up, inspect, sync, or notify the provider channel bound to one goal.
 .TP
+\fBloopx goal\-lifecycle \-\-help\fR
+Preview, stop, or resume a Goal without deleting its history, todos, or evidence.
+.TP
 \fBloopx evidence\-log \-\-goal\-id <goal\-id> \-\-agent\-id <agent\-id> \-\-thin\fR
 Read the current agent's thin public\-safe ledger before replan or handoff.
 .TP
